Worker successfully retrieves and saves commodity listings from the AH.

+++ src/BilgewaterTrade.Core/Models/RealmListing.cs
@@ -7,7 +7,7 @@ public class RealmListing
     public required Item Item { get; set; }
     public int AuctionHouseSnapshotId { get; set; }
     public required AuctionHouseSnapshot AuctionHouseSnapshot { get; set; }
-    public int? BuyoutCopper { get; set; }
+    public long? BuyoutCopper { get; set; }
     public int Quantity {get; set;}
     public Common.TimeLeft TimeLeft { get; set; }
 }
